Update base.rb
Went I make some test I found that if the connection with the client has gone and I want to close it the socket of this connection remains in state ESTABLISHED after the client response a close request. But it cann't the client has gone and it will not be back ;). We need to close the socket in order to avoid DDoS because after some time all sockets will be in use. I use code 1001 to close the socket without sending a close message to client. Following docs I found in https://developer.mozilla.org/es/docs/Web/API/CloseEvent.

If I'm wrong with that solution please tell me. I need to close the socket if client not make to me a ping after X seconds and I cannot find another way to do it.